requirements.txt
pyfiglet==0.8.post1
pydantic==1.8.1
uvicorn==0.13.4
loguru==0.5.3
fastapi==0.63.0
Pillow==8.2.0
aiofiles==0.6.0
slowapi==0.1.4
